project/requirements.txt
beautifulsoup4==4.13.4
duckduckgo_search==8.0.5
fastapi==0.115.14

gradio==5.35.0
ipython==8.30.0
langchain==0.3.26
langchain_chroma==0.2.4
langchain_community==0.3.26
langchain_core==0.3.78
langchain_huggingface==0.3.0
langchain_openai==0.3.34 #0.3.33
langchain_text_splitters==0.3.8
langchain_ollama==0.3.10
langchain_tavily==0.2.11
langchain_deepseek==0.1.4
langgraph==0.6.8
numpy==2.3.1
openai==2.1.0
playwright==1.53.0
pydantic==2.10.3
Pygments==2.19.2
python-dotenv==1.1.1
readability_lxml==0.8.4.1
Requests==2.32.4
networkx==3.2.1
#torch==2.7.1+cu11.8
transformers==4.52.4
sentence-transformers==5.0.0
modelscope==1.27.1
accelerate==1.8.1
sentencepiece==0.2.1
dashscope==1.24.6
#llama-cpp-python==0.3.12
google-search-results==2.4.2
bitsandbytes==0.46.1
pymupdf==1.26.4
rank_bm25==0.2.2
unstructured==0.18.3
markdown==3.4.1
python-docx==1.2.0
trafilatura==2.0.0
langchain-unstructured==0.1.6
libmagic==1.0
python-magic-bin==0.4.14
hf_xet==1.1.10

#triton-window==3.3.1
#环境配置需要注意pydantic的版本